Add files via upload
This simple client/server application follows the following application-layer networking protocol . A
client can send tokens (words) to the server which adds them to a global tokens list in memory which has
a maximum capacity of ten different tokens (there are no duplicates in the list). The tokens list should be
sorted in lexicographical order at any time. There are just three types of requests (SUBMIT, REMOVE and
QUIT). 
The server is be able to handle multiple clients concurrently (simultaneously), using
multithreading (i.e., any number of clients can be simultaneously connected to a single running
instance of the server program and are dealt with by that server at the same time). As specified above,
the server should maintain only a single, global tokens list which is never emptied by the server unless
by means of REMOVE requests (i.e., not a new empty tokens list for each connecting client). That is, all
clients submit their tokens to the same global tokens list on the server, not to different lists.